World international debt EDA
Countries may also take debt to manage its economy. For example, infrastructure spending is one costly ingredient required for a country's citizens to lead comfortable lives. The World Bank is the organization that provides debt to countries. This note book was used to analyze the debt of countries. 
The data is available on world bank